+++ tests/fileReadTests/test1.fasta
