+++ tests/fileReadTests/test2.fasta
