+++ tests/fileReadTests/test3.fasta
