+++ tests/fileReadTests/test4.fasta
