+++ tests/fileReadTests/test5.fasta
