+++ tests/fileReadTests/test6.fasta
