+++ tests/fileReadTests/test7.fasta
