+++ tests/testAssert.h
@@ -10,17 +10,17 @@
 size_t testNum = 0;
 
 void testAssert(bool testSuccess) {
-  if (!testSuccess) {
-    printf("TEST FAIL: Test %zu returned false.\n", testNum);
-  }
-  testNum++;
+    if (!testSuccess) {
+        printf("TEST FAIL: Test %zu returned false.\n", testNum);
+    }
+    testNum++;
 }
 
 void testAssertString(bool testSuccess, char *message) {
-  if (!testSuccess) {
-    printf("TEST FAIL: Test %zu, %s\n", testNum, message);
-  }
-  testNum++;
+    if (!testSuccess) {
+        printf("TEST FAIL: Test %zu, %s\n", testNum, message);
+    }
+    testNum++;
 }
 
 void resetTestNum() { testNum = 0; }
