+++ tests/writeFileTests/statictest.fasta
