+++ tests/writeFileTests/test0.fasta
