+++ tests/writeFileTests/test1.fasta
