+++ tests/writeFileTests/test2.fasta
