+++ tests/writeFileTests/test3.fasta
