+++ tests/writeFileTests/test4.fasta
