+++ tests/writeFileTests/test5.fasta
