+++ tests/writeFileTests/test6.fasta
