+++ tests/writeFileTests/test7.fasta
