+++ tests/writeFileTests/test8.fasta
