+++ tests/writeFileTests/test9.fasta
